[N-09] uint => uint256
Audit Feedback:
>Throughout the codebase, there are uses of `uint` as opposed to `uint256`:
>* In line 40 of `AnvilGovernorDelegator.sol`
>* In line 41 of `AnvilGovernorDelegator.sol`
>* In line 42 of `AnvilGovernorDelegator.sol`
>* In line 310 of `TimeBasedCollateralPool.sol`
>
>In favor of explicitness and style consistency, consider replacing all instances of `uint` with `uint256`.

The AnvilGovernorDelegator.sol instances are no longer applicable after changing it to
an upgradeable proxy (93ebd89411eab300a90f7eceadd45323afe84ce9).

The TimeBasedCollateralPool.sol instance was updated.

+++ contracts/TimeBasedCollateralPool.sol
@@ -308,7 +308,7 @@ contract TimeBasedCollateralPool is ITimeBasedCollateralPool, ICollateralPool, E
         uint256 endOfNextEpoch = getEpochEndTimestamp(currentEpoch + 1);
 
         _claimableCollateral = new ClaimableCollateral[](_tokens.length);
-        for (uint i = 0; i < _tokens.length; i++) {
+        for (uint256 i = 0; i < _tokens.length; i++) {
             address tokenAddress = address(_tokens[i]);
 
             _claimableCollateral[i].endOfCurrentEpochTimestampSeconds = endOfCurrentEpoch;
